LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-stretchy-002-expected.txt


PASS Operator dictionary chunk 1 - stretchy 
FAIL Operator dictionary chunk 2 - stretchy assert_approx_equals: Stretchy property for ⌣ postfix should be 'true' expected 0 +/- 1 but got 42
PASS Operator dictionary chunk 3 - stretchy 
PASS Operator dictionary chunk 4 - stretchy 
FAIL Operator dictionary chunk 5 - stretchy assert_approx_equals: Stretchy property for 𞻰 postfix should be 'true' expected 27.75 +/- 1 but got 66.65625
stretchy for "⌣" (postfix): 
⌣
 VS 
⌣
stretchy for "𞻰" (postfix): 
  
𞻰
 VS 
  
𞻰
